+++ src/net/socket.h
@@ -5,8 +5,10 @@
 #ifndef SMARTCAL_CALDAV_SOCKET_H
 #define SMARTCAL_CALDAV_SOCKET_H
 
+#include <cstdint>
+
 void openMasterSocket();
 
-void openMasterSocket(int port);
+void openMasterSocket(uint_fast16_t port);
 
 #endif //SMARTCAL_CALDAV_SOCKET_H
